requirements.txt
awscli==1.16.152
boto3==1.9.152
djangorestframework==3.9.4
django==2.2.1
django-allauth==0.38.0
django-cors-headers==2.4.0
django-extensions==2.1.6
django-filter==2.1.0
django-ranged-fileresponse==0.1.2
django-storages==1.7.1
django-environ==0.4.5
future==0.16.0
gunicorn==19.9.0
jmespath==0.9.3
psycopg2==2.7.7
python-dateutil==2.6.1
whoosh==2.7.4
zappa==0.47.0